ble_hs_stop.c: Prevent duplicate entries in ble_hs_stop_listeners.
Under circumstances not fully clear sometimes duplicate list elements
are added to the `list ble_hs_stop_listeners`. If that happens, the
function `ble_hs_stop_done()` never terminates.
This PR adds a check into `ble_hs_stop_register_listener()` if
a to-be-added node/sublist already exist in the list and the it does
not add this element. While this is anyhow a good test, it should not
be needed in the MicroPython environment.